Rename HGCN_4_PLS dir to HGScore
