+++ HGScore/casf/ranking_power.py
